src/backend/services/session_manager/requirements.txt
azure-identity==1.20.0
azure-ai-contentsafety==1.0.0
azure-keyvault-secrets==4.7.0
azure-ai-projects==1.0.0b11
Quart==0.19.9
quart-cors==0.7.0
azure-keyvault-secrets==4.7.0
python-dotenv==1.0.0
tenacity==8.4.1
opencensus-ext-azure==1.1.9
opencensus==0.11.2
azure-keyvault==4.2.0
jsonschema[format]==4.19.1
hypercorn==0.14.4
iso8601==2.0.0
aiohttp==3.11.15
pydantic==2.11.9
PyYAML==6.0.1
websockets==13.0.1
aiohttp_cors==0.7.0
redis==5.0.4
setuptools==75.8.0
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
azure-monitor-opentelemetry==1.6.4